Refactored Cilex Console Provider to match Cilex'
I have added the composer.lock so the Service Provider will lock onto a
specific version of its dependencies; newer dependencies may not be
compatible.

The Cilex adapter has been promoted to the top of the Provider's namespace and
the Pimple Service Adapter has been renamed to BaseServiceAdapter. This will put
focus on the Cilex provider and have the Silex provider as adapter to signify its
position as support role.

The Application class has been renamed to ContainerAwareApplication to describe
its function and the setting of the Container has been moved to a separate
setter to  make it uniform with the other ContainerAware* classes of Symfony.

Updated composer.json meta data to match situation and add keywords